requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
faiss-cpu
google-generativeai